Fix vundo-forward
1. Only move once rather than in every step
2. Actually trim undo-list, in previous commit we sometimes don't trim
the undo-list.

Change in looping condition: previously we use node and next, now we
use source, dest, this, and next. When we finish looping, we move from
source to dest. Inside the loop, this is the current node, next is the
next node, and we break the loop if next is nil (early termination) or
step is 0, moved enough steps.

* vundo.el (vundo--trim-undo-list): Add explanation in docstring.
(vundo-forward): Improved the looping condition, now it is easier to
understand. Moves vundo--move-to-node out of the loop. Also fix the
condition for vundo--trim-undo-list, previously NODE could be nil, in
which case the undo-list won't be trimmed.