Fix saved node rendering
The previous approach has a flaw: if a saved node is on the tip of a
branch, it will not be rendered as a saved node. That's because
when we check whether a node A is saved, we check A's first child. But
what if A doesn't have a child (happens when A is on the tip of a
branch)? The right thing is to check the next node in the mod list,
which is what we do in the new version.

* vundo.el (vundo--mod-list-from): Change the local variable SAVED to
CONTAINS-TIMESTAMP for clarity.
(vundo--draw-tree): Check the next mod instead of first child.